requirements.txt
Flask
python-dotenv
requests
urllib3==1.26.6